feat: add use_trait_observe and use_dark_effective
Makes it easy to observe traits, and the example application of this
is use_dark_effective.
