+++ solara/lab/components/__init__.py
@@ -3,4 +3,4 @@ from .confirmation_dialog import ConfirmationDialog  # noqa: F401
 from .input_date import InputDate, InputDateRange  # noqa: F401
 from .menu import ClickMenu, ContextMenu, Menu  # noqa: F401 F403
 from .tabs import Tab, Tabs  # noqa: F401
-from .theming import ThemeToggle, theme  # noqa: F401
+from .theming import ThemeToggle, theme, use_dark_effective  # noqa: F401
