+++ solara/lab/components/theming.py
@@ -11,6 +11,21 @@ theme = Proxy(Theme)
 ipyvuetify.Themes.theme = cast(ipyvuetify.Themes.Theme, theme)
 
 
+def use_dark_effective():
+    """Return True if the frontend is using a dark theme.
+
+    Equivalent of
+
+    ```python
+    solara.use_trait_observe(solara.lab.theme, "dark_effective")
+    ```
+
+    See [use_trait_observe](/api/use_trait_observe).
+
+    """
+    return solara.use_trait_observe(solara.lab.theme, "dark_effective")
+
+
 def _set_theme(themes: Union[Dict[str, Dict[str, str]], None]):
     if themes is None:
         return
