+++ solara/website/pages/api/use_dark_effective.py
@@ -1,0 +1,13 @@
+"""# use_dark_effective
+
+"""
+import solara
+import solara.autorouting
+import solara.lab
+from solara.website.utils import apidoc
+
+from . import NoPage
+
+title = "use_dark_effective"
+Page = NoPage
+__doc__ += apidoc(solara.lab.use_dark_effective)  # type: ignore
